added user.js
