--- src/api/users.js
@@ -1,90 +1,0 @@
-const usersRouter = require('express').Router();
-const jwt = require('jsonwebtoken')
-
-const {createUser, getByUsername, getAllUsers, getUserById, getUserByUsername} = require ('../db')
-const SECRET = require('./secret')
-
-usersRouter.post('/users', async (req, res, next) => {
-    try {
-        console.log('CREATING THE USER');
-        if(req.body.password.length < 8) {
-            console.error('password too short')
-            next()
-            if(req.body.username === username) {
-                console.log('name already taken')
-                next()
-            }
-        }
-            const user = await createUser(req.body);
-            console.log(user)
-            res.send({user});
-    } catch (error) {
-        console.log('THERE WAS AN ERROR');
-        next(error);
-    }
-  });
-
-  usersRouter.post('/users', async (req, res, next) => {
-    try {
-        const user = await getByUsername
-        (req.body.username, req.body.password, req.body.firstname, req.body.lastname, req.body.location);
-        const token = jwt.sign({
-            id: user.id,
-            username: user.username
-            
-        }, SECRET);
-        console.log(token);
-        res.send({token});
-    } catch (error) {
-        next(error);
-    }
-});
-
-usersRouter.post('/users', async (req, res, next) => {
-    try {
-        const user = await getAllUsers
-        (req.body.username, req.body.password, req.body.firstname, req.body.lastname, req.body.location);
-        const token = jwt.sign({
-            id: user.id,
-            username: user.username
-        }, SECRET);
-        console.log(token);
-        res.send({token});
-    } catch (error) {
-        next(error);
-    }
-});
-
-usersRouter.post('/users', async (req, res, next) => {
-    try {
-        const user = await getUserById
-        (req.body.username, req.body.password, req.body.firstname, req.body.lastname, req.body.location);
-        const token = jwt.sign({
-            id: user.id,
-            username: user.username
-        }, SECRET);
-        console.log(token);
-        res.send({token});
-    } catch (error) {
-        next(error);
-    }
-});
-
-usersRouter.post('/users', async (req, res, next) => {
-    try {
-        const user = await getUserByUsername
-        (req.body.username, req.body.password, req.body.firstname, req.body.lastname, req.body.location);
-        const token = jwt.sign({
-            id: user.id,
-            username: user.username
-        }, SECRET);
-        console.log(token);
-        res.send({token});
-    } catch (error) {
-        next(error);
-    }
-});
-
-
-
-module.exports = usersRouter
